enowars/services/3xam/backend_internal/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
packaging==20.9
pluggy==0.13.1
protobuf==3.16.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
dissononce
gunicorn==20.1.0
requests==2.25.1
psutil==5.8.0
gevent==21.1.2
